+++ plugin/evm/validators/interfaces/interfaces.go
@@ -13,7 +13,9 @@ import (
 )
 
 type ValidatorReader interface {
-	// GetValidatorAndUptime returns the uptime of the validator specified by validationID
+	// GetValidatorAndUptime returns the calculated uptime of the validator specified by validationID
+	// and the last updated time.
+	// GetValidatorAndUptime holds the chain context lock while performing the operation and can be called concurrently.
 	GetValidatorAndUptime(validationID ids.ID) (stateinterfaces.Validator, time.Duration, time.Time, error)
 }
 
@@ -21,10 +23,15 @@ type Manager interface {
 	stateinterfaces.State
 	avalancheuptime.Manager
 	ValidatorReader
-
-	// Sync updates the validator set managed
-	// by the manager
-	Sync(ctx context.Context) error
+	// Initialize initializes the validator manager
+	// by syncing the validator state with the current validator set
+	// and starting the uptime tracking.
+	// Initialize holds the chain context lock while performing the operation.
+	Initialize(ctx context.Context) error
+	// Shutdown stops the uptime tracking and writes the validator state to the database.
+	// Shutdown holds the chain context lock while performing the operation.
+	Shutdown() error
 	// DispatchSync starts the sync process
+	// DispatchSync holds the chain context lock while performing the sync.
 	DispatchSync(ctx context.Context)
 }