+++ RELEASES.md
@@ -5,6 +5,7 @@
 * Bump golangci-lint to v1.63 and add linters
 * Refactored trie_prefetcher.go to be structurally similar to upstream.
 * Remove legacy gossip handler and metrics
+* Fixed concurrency issue in validators/uptime manager
 
 ## [v0.7.0](https://github.com/ava-labs/subnet-evm/releases/tag/v0.7.0)